Seg ok %Lf ok

+++ src/process/conversion_id/process_f.c
@@ -6,7 +6,7 @@
 /*   By: jfleury <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/14 14:30:04 by jfleury           #+#    #+#             */
-/*   Updated: 2019/01/25 14:56:10 by jfleury          ###   ########.fr       */
+/*   Updated: 2019/01/28 10:43:01 by jfleury          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@ char			*ft_process_f(t_conv_spec conv_spec, va_list *ap)
 	if (conv_spec.modifier[3] == 1)
 		return (ft_itoa_double(conv_spec, va_arg(*ap, double)));
 	if (conv_spec.modifier[4] == 1)
-		return (ft_itoa_double(conv_spec, va_arg(*ap, long double)));
+		return (ft_itoa_double(conv_spec, va_arg(*ap, double)));
 	return (ft_itoa_double(conv_spec, va_arg(*ap, double)));
 }
